feat: 設定の項目を移動
