+++ modules/features/setting/src/main/java/net/pantasystem/milktea/setting/activities/SettingAppearanceActivity.kt
@@ -19,6 +19,7 @@ import androidx.compose.material.icons.filled.Remove
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.lifecycleScope
 import coil.compose.rememberAsyncImagePainter
@@ -221,6 +222,31 @@ class SettingAppearanceActivity : AppCompatActivity() {
                             }
                         }
                         SettingSection(title = stringResource(id = R.string.settings_note)) {
+                            Column(Modifier.padding(horizontal = 16.dp)) {
+                                Text(text = stringResource(id = R.string.auto_note_folding), fontWeight = FontWeight.Bold)
+                                Text(text = stringResource(id = R.string.height_limit))
+                                Slider(
+                                    value = currentConfigState.noteExpandedHeightSize.let {
+                                        val v =
+                                            currentConfigState.noteExpandedHeightSize.toFloat() / 1000f
+                                        if (it in 0..1000) {
+                                            v
+                                        } else {
+                                            1f
+                                        }
+                                    },
+                                    onValueChange = {
+                                        val v = (it * 1000f).toInt()
+                                        currentConfigState = currentConfigState.copy(
+                                            noteExpandedHeightSize = if (v > 50) {
+                                                v
+                                            } else 50
+                                        )
+
+                                    },
+                                    modifier = Modifier.fillMaxWidth(),
+                                )
+                            }
                             Text(
                                 stringResource(id = R.string.note_opacity),
                                 modifier = Modifier.padding(horizontal = 16.dp)
